kb-kafka-pipeline/db_sync/pyproject.toml
[tool.poetry]
name = "db_sync"
version = "0.1.0"
description = ""
authors = ["JusbrasilDevs"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
aiokafka = "^0.11.0"
psycopg2-binary = "^2.9.9"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"

[tool.poetry.dev-dependencies]


[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
